% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-screenshot.R
\name{generate_screenshot}
\alias{generate_screenshot}
\title{Generate a webpage screenshot}
\usage{
generate_screenshot(
  url,
  quality = 100L,
  full_page = FALSE,
  mobile = FALSE,
  width = NULL,
  type = c("jpeg", "png", "webp"),
  ...
)
}
\arguments{
\item{url}{the URL to screenshot}

\item{quality}{image quality. Defaults to 100 (i.e. 100\%)}

\item{full_page}{generate a full page scren shot (may be very "long"). Default \code{FALSE}.}

\item{mobile}{generate a screenshot of a mobile version of the website? Default \code{FALSE}.}

\item{width}{screen width (pixels). Defaults to \code{NULL} which uses the Statically
default (which appears to be 1280).}

\item{type}{which image format to request. This function returns a \code{{magick}} object
regardless of what is specified here. Using a type other than "\code{jpeg}"
(the API default and this function's default) may result in subtly different
captures.}

\item{...}{passed on to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
\code{{magick}} object
}
\description{
Generate a webpage screenshot
}
\examples{
generate_screenshot("https://cran.r-project.org")
}
