% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statically-package.R
\docType{package}
\name{statically}
\alias{statically}
\alias{statically-package}
\title{Generate Webpage Screenshots Using the Statically API}
\description{
Statically (\url{https://statically.io/screenshot/}) maintains a free
API to generate screenshots of webpages. Tools are provided to capture
website screenshots using this API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/statically}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/statically/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
