% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statically-package.R
\docType{package}
\name{statically}
\alias{statically}
\alias{statically-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/statically}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/statically/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
